added forecast list screen

+++ app/src/main/java/com/tamarcusg/openweather/repository/model/Forecast.kt
@@ -1,7 +1,7 @@
 package com.tamarcusg.openweather.repository.model
 
 internal data class Forecast(
-    val temperature: Double,
-    val feelsLike: Double,
+    val temperature: Int,
+    val feelsLike: Int,
     val weatherDescriptions: List<WeatherDescription>
 )
